Create Bisection Algorithm
Bisection Algorithm is an algorithm used to approximate the root of a given continuous function. First we define an interval (a,b) where f(a)<0 and f(b)>0. Since the function is continuous, we know there has to be a root between a and b. For our first root approximation, we define a new value c for (a+b)/2.