Create Newton Raphson Algorithm
Newton-Raphson Algorithm is an open method used in the process  of root finding operations. The main idea is that for an initial approximation of x1 , we draw the tangent line to our function at x1. The point where our tangent line crosses the x-axis becomes our next approximation. We repeat this process until our error rate condition is satisfied. The mathematical model of Newton-Raphson Algorithm is as follows: xi+1 = xi - f(xi)/f'(xi+1) In order to get this equation, we express the slope of the tangent line in two different ways : as the tangent of the angle between the line and the axis and as the derivative of the function at that point.